# tvsos-backend/swagger-docs/swagger.yaml
definitions:
  model.Poi:
    properties:
      id:
        type: integer
      lat:
        type: number
      lon:
        type: number
      name:
        type: string
      status:
        type: integer
      tybe:
        type: integer
    type: object
  model.Vehicle:
    properties:
      id:
        type: integer
      lat:
        type: number
      lon:
        type: number
      speed:
        type: number
      status:
        type: integer
      updateTime:
        type: string
    type: object
  result.Result:
    properties:
      code:
        description: 1表示业务成功 2表示业务失败
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  title: Mover API
  version: "1.0"
paths:
  /pois:
    get:
      consumes:
      - application/json
      parameters:
      - description: poi 名称
        in: query
        name: name
        type: string
      - description: poi 类型
        in: query
        name: tybe
        type: integer
      - description: poi 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/result.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Poi'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/result.Result'
      summary: 筛选/获取poi列表
      tags:
      - Poi
  /pois/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/result.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.Poi'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/result.Result'
      summary: 根据id获取poi信息
      tags:
      - Poi
  /vehicles:
    get:
      consumes:
      - application/json
      parameters:
      - description: vehicle 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/result.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Vehicle'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/result.Result'
      summary: 查询/筛选车辆列表
      tags:
      - Vehicle
swagger: "2.0"